Additions/linux/vboxvideo: Fix reporting invalid suggested-offset-properties (Partial commit of kernel change.) bugref:8282: Additions/linux: track kernel changes to vboxvideo in our own tree
    The x and y hints receives from the host are unsigned 32 bit integers and
    they get set to -1 (0xffffffff) when invalid. Before this commit the
    vboxvideo driver was storing them in an u16 causing the -1 to be truncated
    to 65535 which, once reported to userspace, was breaking gnome 3.26+
    in Wayland mode.
    
    This commit stores the host values in 32 bit variables, removing the
    truncation and checks for -1, replacing it with 0 as -1 is not a valid
    suggested-offset-property value. Likewise the properties are now
    initialized to 0 instead of -1, since -1 is not a valid value.
    This fixes gnome 3.26+ in Wayland mode not working with the vboxvideo
    driver.
    
    Reported-by: Gianfranco Costamagna <locutusofborg@debian.org>
    Cc: stable@vger.kernel.org
    Cc: Michael Thayer <michael.thayer@oracle.com>

svn:sync-xref-src-repo-rev: r128314

+++ src/VBox/Additions/linux/drm/vbox_drv.h
@@ -1,4 +1,4 @@
-/* $Id: vbox_drv.h 76563 2019-01-01 03:53:56Z knut.osmundsen@oracle.com $ */
+/* $Id: vbox_drv.h 76937 2019-01-22 16:39:05Z noreply@oracle.com $ */
 /** @file
  * VirtualBox Additions Linux kernel video driver
  */
@@ -236,8 +236,8 @@ struct vbox_connector {
 	char name[32];
 	struct vbox_crtc *vbox_crtc;
 	struct {
-		u16 width;
-		u16 height;
+		u32 width;
+		u32 height;
 		bool disconnected;
 	} mode_hint;
 };
